requirements.txt
fastapi==0.70.0google-api-python-client==2.30.0google-auth-oauthlib==0.4.6pyyaml==6.0uvicorn==0.15.0